Fix error from the MaterialFox recipe

+++ galactic-aurora/recipes/materialfox/install.sh
@@ -5,8 +5,14 @@ rice_cooker_recipe_cache_dir
 rice_cooker_debug "Downloading git repository"
 git clone git@github.com:muckSponge/MaterialFox.git ${RECIPE_CACHE_DIR}
 
+rice_cooker_debug "Attempting to switch to the beta branch"
+cd ${RECIPE_CACHE_DIR} && git checkout beta
+
+if [ -d "${RECIPE_DIST_DIR}" ]; then
+    rm -r ${RECIPE_DIST_DIR}
+fi
+
 rice_cooker_debug "Migrating out of cache"
-rm -r ${RECIPE_DIST_DIR}
 cp -r ${RECIPE_CACHE_DIR}/chrome ${RECIPE_DIST_DIR}
 
 rice_cooker_debug "Replacing darkmode variables"
